+++ src/demo/atlas-id.ts
@@ -4,11 +4,16 @@
  * animation duration, from the Atlas. The tag convention used here is
  * <file stem>-<state>.
  */
-export type AtlasID = typeof AtlasID[keyof typeof AtlasID]
-export const AtlasID = <const>{
-  FrogEat: 'frog-eat',
-  FrogIdle: 'frog-idle',
-  BackpackerWalkRight: 'backpacker-walkRight',
-  BackpackerWalkDown: 'backpacker-walkDown',
-  BackpackerWalkUp: 'backpacker-walkUp'
+export type AtlasID = Parameters<typeof AtlasID.values['has']>[0]
+
+export namespace AtlasID {
+  export const values = Object.freeze(
+    new Set(<const>[
+      'frog-eat',
+      'frog-idle',
+      'backpacker-walkRight',
+      'backpacker-walkDown',
+      'backpacker-walkUp'
+    ])
+  )
 }